Rewrite "unicode friendly" regex characters to ascii versions if unicode support is disabled
Many of the unicode issues we have been dealing with come from users who
don't actually need unicode support, but they use \s or \w to match
spaces or words.  The user expectation here may be that these will match
ascii whitespace or ascii words, but the regex library requires unicode
support to handle them.  We've previously updated the docs to point this
out, but the more ergonomic thing to do might be to just support these
as ascii matches by rewriting them in lalrpop before we pass them on to
regex.