Only generate unicode tokens if the unicode feature in enabled
regex_syntax::parse() converts our regex strings into the HIR of the
regex, part of that includes unpacking various metacharacters into a
list of symbols.  In many cases, this expansion changes depending if it
should expand into unicode or not.

Prior to #814, we were still outputting unicode regexes unconditionally,
but regex internals seem to be compiling them away and avoiding errors.
The switch in #814 caused these to be result in real errors.

Follow-up work will be needed to determine why existing tests didn't
detect this.