Add benchmarking support using criterion
This enables us to benchmark proposed changes directly from lalrpop to
check both their impact on the time taken to generate code and the
performance of generated code.

This is intended as a starting point to build on and add additional
benchmarking tests as we move forward, so it only has one compile-time
test and one generated code test so far.  With the basic infrastructure
in place adding additional tests should be straightforward.